# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación API",
        "description": "Documentación API",
        "contact": {
            "name": "API Support",
            "url": "http://www.fmt.com.do",
            "email": "ivan.firestone@fmt.com.do"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/usuarios": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "description": "Listado de Usuarios",
                "operationId": "user_index",
                "parameters": [
                    {
                        "name": "busqueda",
                        "in": "query",
                        "description": "Parametro de busqueda",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "string",
                            "example": "admin"
                        }
                    },
                    {
                        "name": "estado",
                        "in": "query",
                        "description": "Filtra los usuarios entre activos y no activos",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "activos | inactivos",
                            "example": "activos"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "Json",
                                            "example": "[...]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/usuarios/store": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "description": "Crear un nuevo usuario",
                "operationId": "user_store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "email",
                                    "username",
                                    "role",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "nombre": {
                                        "description": "Nombre del usuario",
                                        "type": "string",
                                        "example": "Juan Perez"
                                    },
                                    "email": {
                                        "description": "Email del usuario",
                                        "type": "string",
                                        "example": "user@email.aqui"
                                    },
                                    "username": {
                                        "description": "Nombre de Usuario",
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "role": {
                                        "description": "Model Id",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "telefono": {
                                        "description": "Telefono Contacto",
                                        "type": "string",
                                        "example": "(999) 555-5555"
                                    },
                                    "movil": {
                                        "description": "Movil Contacto",
                                        "type": "string",
                                        "example": "(999) 444-3333"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string",
                                        "example": "SuperStrongPasswordHere"
                                    },
                                    "password_confirmation": {
                                        "description": "Password Confirmation",
                                        "type": "string",
                                        "example": "SuperStrongPasswordHere"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Stored",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "Json",
                                            "example": "[...]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/usuarios/{usuario}/show": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "description": "Mostrar un usuario",
                "operationId": "user_show",
                "parameters": [
                    {
                        "name": "usuario",
                        "in": "path",
                        "description": "Model Id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "Json",
                                            "example": "[...]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/usuarios/{usuario}/update": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "description": "Actualizar un usuario",
                "operationId": "user_update",
                "parameters": [
                    {
                        "name": "usuario",
                        "in": "path",
                        "description": "Model Id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "integer",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "username",
                                    "email",
                                    "role"
                                ],
                                "properties": {
                                    "nombre": {
                                        "description": "Nombre del usuario",
                                        "type": "string",
                                        "example": "Juan Perez"
                                    },
                                    "email": {
                                        "description": "Email del usuario",
                                        "type": "string",
                                        "example": "user@email.aqui"
                                    },
                                    "username": {
                                        "description": "Nombre de Usuario",
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "role": {
                                        "description": "Role ID Asignado",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "telefono": {
                                        "description": "Telefono Contacto",
                                        "type": "string",
                                        "example": "(999) 555-5555"
                                    },
                                    "movil": {
                                        "description": "Movil Contacto",
                                        "type": "string",
                                        "example": "(999) 444-3333"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Stored",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "Json",
                                            "example": "[...]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    }
}